Moved proxy code into "proxy" package, proxy can now be started and stopped in a non-blocking way
